+++ builder.go
@@ -34,7 +34,10 @@ const (
 	// FieldTypeString represents N registers as string value. Use `Field.Length` to length of string.
 	FieldTypeString FieldType = 13
 
-	maxFieldTypeValue = uint8(13)
+	// FieldTypeCoil represents single discrete/coil value (used by FC1/FC2).
+	FieldTypeCoil FieldType = 14
+
+	maxFieldTypeValue = uint8(14)
 )
 
 // FieldType is enum type for data types that Field can represent
@@ -46,17 +49,19 @@ type Fields []Field
 // Field is distinct field be requested and extracted from response
 // Tag `mapstructure` allows you to marshal https://github.com/spf13/viper supported configuration format to the Field
 type Field struct {
+	Name string `json:"Name" mapstructure:"Name"`
+
 	ServerAddress string `json:"server_address" mapstructure:"server_address"` // [network://]host:port
 	UnitID        uint8  `json:"unit_id" mapstructure:"unit_id"`
+	// Address of the register (first register of that data type) or discrete/coil address in modbus. Addresses are 0-based.
+	Address uint16    `json:"address" mapstructure:"address"`
+	Type    FieldType `json:"type" mapstructure:"type"`
 
-	RegisterAddress uint16    `json:"register_address" mapstructure:"register_address"`
-	Type            FieldType `json:"type" mapstructure:"type"`
-	Bit             uint8     `json:"bit" mapstructure:"bit"`
-	FromHighByte    bool      `json:"from_high_byte" mapstructure:"from_high_byte"`
-	Length          uint8     `json:"Length" mapstructure:"Length"`
-
-	ByteOrder packet.ByteOrder `json:"byte_order" mapstructure:"byte_order"`
-	Name      string           `json:"Name" mapstructure:"Name"`
+	// Only relevant to register function fields
+	Bit          uint8            `json:"bit" mapstructure:"bit"`
+	FromHighByte bool             `json:"from_high_byte" mapstructure:"from_high_byte"`
+	Length       uint8            `json:"Length" mapstructure:"Length"`
+	ByteOrder    packet.ByteOrder `json:"byte_order" mapstructure:"byte_order"`
 }
 
 // registerSize returns how many register/words does this field would take in modbus response
@@ -77,7 +82,7 @@ func (f *Field) registerSize() uint16 {
 }
 
 // Validate checks if Field is values are correctly filled
-func (f Field) Validate() error {
+func (f *Field) Validate() error {
 	if f.ServerAddress == "" {
 		return errors.New("field server address can not be empty")
 	}
@@ -97,34 +102,34 @@ func (f Field) Validate() error {
 }
 
 // ExtractFrom extracts field value from given registers data
-func (f Field) ExtractFrom(registers *packet.Registers) (interface{}, error) {
+func (f *Field) ExtractFrom(registers *packet.Registers) (interface{}, error) {
 	switch f.Type {
 	case FieldTypeBit:
-		return registers.Bit(f.RegisterAddress, f.Bit)
+		return registers.Bit(f.Address, f.Bit)
 	case FieldTypeByte:
-		return registers.Byte(f.RegisterAddress, f.FromHighByte)
+		return registers.Byte(f.Address, f.FromHighByte)
 	case FieldTypeUint8:
-		return registers.Uint8(f.RegisterAddress, f.FromHighByte)
+		return registers.Uint8(f.Address, f.FromHighByte)
 	case FieldTypeInt8:
-		return registers.Int8(f.RegisterAddress, f.FromHighByte)
+		return registers.Int8(f.Address, f.FromHighByte)
 	case FieldTypeUint16:
-		return registers.Uint16(f.RegisterAddress)
+		return registers.Uint16(f.Address)
 	case FieldTypeInt16:
-		return registers.Int16(f.RegisterAddress)
+		return registers.Int16(f.Address)
 	case FieldTypeUint32:
-		return registers.Uint32WithByteOrder(f.RegisterAddress, f.ByteOrder)
+		return registers.Uint32WithByteOrder(f.Address, f.ByteOrder)
 	case FieldTypeInt32:
-		return registers.Int32WithByteOrder(f.RegisterAddress, f.ByteOrder)
+		return registers.Int32WithByteOrder(f.Address, f.ByteOrder)
 	case FieldTypeUint64:
-		return registers.Uint64WithByteOrder(f.RegisterAddress, f.ByteOrder)
+		return registers.Uint64WithByteOrder(f.Address, f.ByteOrder)
 	case FieldTypeInt64:
-		return registers.Int64WithByteOrder(f.RegisterAddress, f.ByteOrder)
+		return registers.Int64WithByteOrder(f.Address, f.ByteOrder)
 	case FieldTypeFloat32:
-		return registers.Float32WithByteOrder(f.RegisterAddress, f.ByteOrder)
+		return registers.Float32WithByteOrder(f.Address, f.ByteOrder)
 	case FieldTypeFloat64:
-		return registers.Float64WithByteOrder(f.RegisterAddress, f.ByteOrder)
+		return registers.Float64WithByteOrder(f.Address, f.ByteOrder)
 	case FieldTypeString:
-		return registers.StringWithByteOrder(f.RegisterAddress, f.Length, f.ByteOrder)
+		return registers.StringWithByteOrder(f.Address, f.Length, f.ByteOrder)
 	}
 	return nil, errors.New("extraction failure due unknown field type")
 }
@@ -196,8 +201,21 @@ func (b *Builder) Bit(registerAddress uint16, bit uint8) *BField {
 			UnitID:        b.unitID,
 			Type:          FieldTypeBit,
 
-			RegisterAddress: registerAddress,
-			Bit:             bit,
+			Address: registerAddress,
+			Bit:     bit,
+		},
+	}
+}
+
+// Coil adds discrete/coil field to Builder to be requested and extracted by FC1/FC2.
+func (b *Builder) Coil(address uint16) *BField {
+	return &BField{
+		Field{
+			ServerAddress: b.serverAddress,
+			UnitID:        b.unitID,
+			Type:          FieldTypeCoil,
+
+			Address: address,
 		},
 	}
 }
@@ -210,8 +228,8 @@ func (b *Builder) Byte(registerAddress uint16, fromHighByte bool) *BField {
 			UnitID:        b.unitID,
 			Type:          FieldTypeByte,
 
-			RegisterAddress: registerAddress,
-			FromHighByte:    fromHighByte,
+			Address:      registerAddress,
+			FromHighByte: fromHighByte,
 		},
 	}
 }
@@ -224,8 +242,8 @@ func (b *Builder) Uint8(registerAddress uint16, fromHighByte bool) *BField {
 			UnitID:        b.unitID,
 			Type:          FieldTypeUint8,
 
-			RegisterAddress: registerAddress,
-			FromHighByte:    fromHighByte,
+			Address:      registerAddress,
+			FromHighByte: fromHighByte,
 		},
 	}
 }
@@ -238,8 +256,8 @@ func (b *Builder) Int8(registerAddress uint16, fromHighByte bool) *BField {
 			UnitID:        b.unitID,
 			Type:          FieldTypeInt8,
 
-			RegisterAddress: registerAddress,
-			FromHighByte:    fromHighByte,
+			Address:      registerAddress,
+			FromHighByte: fromHighByte,
 		},
 	}
 }
@@ -252,7 +270,7 @@ func (b *Builder) Uint16(registerAddress uint16) *BField {
 			UnitID:        b.unitID,
 			Type:          FieldTypeUint16,
 
-			RegisterAddress: registerAddress,
+			Address: registerAddress,
 		},
 	}
 }
@@ -265,7 +283,7 @@ func (b *Builder) Int16(registerAddress uint16) *BField {
 			UnitID:        b.unitID,
 			Type:          FieldTypeInt16,
 
-			RegisterAddress: registerAddress,
+			Address: registerAddress,
 		},
 	}
 }
@@ -278,7 +296,7 @@ func (b *Builder) Uint32(registerAddress uint16) *BField {
 			UnitID:        b.unitID,
 			Type:          FieldTypeUint32,
 
-			RegisterAddress: registerAddress,
+			Address: registerAddress,
 		},
 	}
 }
@@ -291,7 +309,7 @@ func (b *Builder) Int32(registerAddress uint16) *BField {
 			UnitID:        b.unitID,
 			Type:          FieldTypeInt32,
 
-			RegisterAddress: registerAddress,
+			Address: registerAddress,
 		},
 	}
 }
@@ -304,7 +322,7 @@ func (b *Builder) Uint64(registerAddress uint16) *BField {
 			UnitID:        b.unitID,
 			Type:          FieldTypeUint64,
 
-			RegisterAddress: registerAddress,
+			Address: registerAddress,
 		},
 	}
 }
@@ -317,7 +335,7 @@ func (b *Builder) Int64(registerAddress uint16) *BField {
 			UnitID:        b.unitID,
 			Type:          FieldTypeInt64,
 
-			RegisterAddress: registerAddress,
+			Address: registerAddress,
 		},
 	}
 }
@@ -330,7 +348,7 @@ func (b *Builder) Float32(registerAddress uint16) *BField {
 			UnitID:        b.unitID,
 			Type:          FieldTypeFloat32,
 
-			RegisterAddress: registerAddress,
+			Address: registerAddress,
 		},
 	}
 }
@@ -343,7 +361,7 @@ func (b *Builder) Float64(registerAddress uint16) *BField {
 			UnitID:        b.unitID,
 			Type:          FieldTypeFloat64,
 
-			RegisterAddress: registerAddress,
+			Address: registerAddress,
 		},
 	}
 }
@@ -357,13 +375,13 @@ func (b *Builder) String(registerAddress uint16, length uint8) *BField {
 			Type:          FieldTypeString,
 			Length:        length,
 
-			RegisterAddress: registerAddress,
+			Address: registerAddress,
 		},
 	}
 }
 
-// RegisterRequest helps to connect requested fields to responses
-type RegisterRequest struct {
+// BuilderRequest helps to connect requested fields to responses
+type BuilderRequest struct {
 	packet.Request
 
 	// ServerAddress is modbus server address where request should be sent
@@ -383,8 +401,14 @@ type RegistersResponse interface {
 	AsRegisters(requestStartAddress uint16) (*packet.Registers, error)
 }
 
+// CoilsResponse is marker interface for responses returning coil/discrete data
+type CoilsResponse interface {
+	packet.Response
+	IsCoilSet(startAddress uint16, coilAddress uint16) (bool, error)
+}
+
 // AsRegisters returns response data as Register to more convenient access
-func (r RegisterRequest) AsRegisters(response RegistersResponse) (*packet.Registers, error) {
+func (r BuilderRequest) AsRegisters(response RegistersResponse) (*packet.Registers, error) {
 	return response.AsRegisters(r.StartAddress)
 }
 
@@ -401,7 +425,17 @@ var ErrorFieldExtractHadError = errors.New("field extraction had an error. check
 // ExtractFields extracts Field values from given response. When continueOnExtractionErrors is true and error occurs
 // during extraction, this method does not end but continues to extract all Fields and returns ErrorFieldExtractHadError
 // at the end. To distinguish errors check FieldValue.Error field.
-func (r RegisterRequest) ExtractFields(response RegistersResponse, continueOnExtractionErrors bool) ([]FieldValue, error) {
+func (r BuilderRequest) ExtractFields(response packet.Response, continueOnExtractionErrors bool) ([]FieldValue, error) {
+	switch resp := response.(type) {
+	case RegistersResponse:
+		return r.extractRegisterFields(resp, continueOnExtractionErrors)
+	case CoilsResponse:
+		return r.extractCoilFields(resp, continueOnExtractionErrors)
+	}
+	return nil, errors.New("can not extract fields from unsupported response type")
+}
+
+func (r BuilderRequest) extractRegisterFields(response RegistersResponse, continueOnExtractionErrors bool) ([]FieldValue, error) {
 	regs, err := response.AsRegisters(r.StartAddress)
 	if err != nil {
 		return nil, err
@@ -434,22 +468,71 @@ func (r RegisterRequest) ExtractFields(response RegistersResponse, continueOnExt
 	return result, nil
 }
 
+func (r BuilderRequest) extractCoilFields(response CoilsResponse, continueOnExtractionErrors bool) ([]FieldValue, error) {
+	hadErrors := false
+	capacity := 0
+	if continueOnExtractionErrors {
+		capacity = len(r.Fields)
+	}
+	result := make([]FieldValue, 0, capacity)
+	for _, f := range r.Fields {
+		vTmp, err := response.IsCoilSet(r.StartAddress, f.Address)
+
+		if err != nil && !continueOnExtractionErrors {
+			return nil, fmt.Errorf("field extraction failed. name: %v err: %w", f.Name, err)
+		}
+		if !hadErrors && err != nil {
+			hadErrors = true
+		}
+		tmp := FieldValue{
+			Field: f,
+			Value: vTmp,
+			Error: err,
+		}
+		result = append(result, tmp)
+	}
+	if hadErrors {
+		return result, ErrorFieldExtractHadError
+	}
+	return result, nil
+}
+
 // ReadHoldingRegistersTCP combines fields into TCP Read Holding Registers (FC3) requests
-func (b *Builder) ReadHoldingRegistersTCP() ([]RegisterRequest, error) {
-	return split(b.fields, "fc3_tcp")
+func (b *Builder) ReadHoldingRegistersTCP() ([]BuilderRequest, error) {
+	return split(b.fields, splitToFC3TCP)
 }
 
 // ReadHoldingRegistersRTU combines fields into RTU Read Holding Registers (FC3) requests
-func (b *Builder) ReadHoldingRegistersRTU() ([]RegisterRequest, error) {
-	return split(b.fields, "fc3_rtu")
+func (b *Builder) ReadHoldingRegistersRTU() ([]BuilderRequest, error) {
+	return split(b.fields, splitToFC3RTU)
 }
 
 // ReadInputRegistersTCP combines fields into TCP Read Input Registers (FC4) requests
-func (b *Builder) ReadInputRegistersTCP() ([]RegisterRequest, error) {
-	return split(b.fields, "fc4_tcp")
+func (b *Builder) ReadInputRegistersTCP() ([]BuilderRequest, error) {
+	return split(b.fields, splitToFC4TCP)
 }
 
 // ReadInputRegistersRTU combines fields into RTU Read Input Registers (FC4) requests
-func (b *Builder) ReadInputRegistersRTU() ([]RegisterRequest, error) {
-	return split(b.fields, "fc4_rtu")
+func (b *Builder) ReadInputRegistersRTU() ([]BuilderRequest, error) {
+	return split(b.fields, splitToFC4RTU)
+}
+
+// ReadCoilsTCP combines fields into TCP Read Coils (FC1) requests
+func (b *Builder) ReadCoilsTCP() ([]BuilderRequest, error) {
+	return split(b.fields, splitToFC1TCP)
+}
+
+// ReadCoilsRTU combines fields into RTU Read Coils (FC1) requests
+func (b *Builder) ReadCoilsRTU() ([]BuilderRequest, error) {
+	return split(b.fields, splitToFC1RTU)
+}
+
+// ReadDiscreteInputsTCP combines fields into TCP Read Discrete Inputs (FC2) requests
+func (b *Builder) ReadDiscreteInputsTCP() ([]BuilderRequest, error) {
+	return split(b.fields, splitToFC2TCP)
+}
+
+// ReadDiscreteInputsRTU combines fields into RTU Read Discrete Inputs (FC2) requests
+func (b *Builder) ReadDiscreteInputsRTU() ([]BuilderRequest, error) {
+	return split(b.fields, splitToFC2RTU)
 }